+++ libnopegl/src/image.h
@@ -66,7 +66,7 @@ struct image_params {
 struct image {
     struct image_params params;
     struct texture *planes[4];
-    int nb_planes;
+    size_t nb_planes;
     NGLI_ALIGNED_MAT(color_matrix);
     /* mutable fields after initialization */
     NGLI_ALIGNED_MAT(coordinates_matrix);
